requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
decorator==5.1.1
Django==3.2
django-cleanup==7.0.0
django-rest-knox==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.12.0
drf-yasg2==1.19.4
dynaconf==3.1.12
hvac==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
knox==0.1.14
loguru==0.7.0
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
pycparser==2.21
pyhcl==0.4.4
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
shortuuid==1.0.11
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.2
validators==0.20.0
